shuffling around command names
