+++ neuroscout_cli/cli.py
@@ -3,8 +3,8 @@ neuroscout
 
 Usage:
     neuroscout run [first_level|group_level] [-w <work_dir>|-c|--jobs=<n>] <bundle_id>
-    neuroscout get [bundle|data] <bundle_id>
-    neuroscout clone <bundle_id>
+    neuroscout install [bundle|data] <bundle_id>
+    neuroscout get <bundle_id>
     neuroscout -h | --help
     neuroscout --version
