--- neuroscout_cli/commands/clone.py
@@ -1,9 +1,0 @@
-from neuroscout_cli.commands.base import Command
-
-
-class Clone(Command):
-
-    ''' Command for listing available files. '''
-
-    def run(self):
-        pass
